added some data

+++ Server/consumables/migrations/0001_initial.py
@@ -14,7 +14,7 @@ class Migration(migrations.Migration):
             fields=[
                 ('id', models.BigAutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
                 ('name', models.CharField(max_length=100)),
-                ('price', models.IntegerField()),
+                ('price', models.DecimalField(max_digits=5, decimal_places=2)),
                 ('description', models.TextField()),
                 ('photo_main', models.ImageField(upload_to='photos/%Y/%m/%d/')),
                 ('cuisine_type', models.CharField(choices=[('FD', 'Food'), ('DR', 'Drink')], max_length=2)),
